Add tests and move script
