+++ packages/semantic-tokens/.eslintrc.json
@@ -1,5 +1,12 @@
 {
   "extends": ["plugin:@fluentui/eslint-plugin/react"],
   "root": true,
-  "rules": {}
+  "rules": {
+    "no-restricted-imports": [
+      "error",
+      {
+        "patterns": ["react", "react-dom"]
+      }
+    ]
+  }
 }
