+++ packages/semantic-tokens/utils/escapeInlineToken.test.ts
@@ -1,4 +1,5 @@
 import { escapeInlineToken } from './escapeInlineToken';
+
 describe('escapeInlineToken', () => {
   it('Handles a variable replacement string as pure text', () => {
     expect(escapeInlineToken('textGlobalDisplay1FontsizeRaw')).toMatch('${textGlobalDisplay1FontsizeRaw}');
